Moved card reader tracker to the payments package
